#4484 python int type hint was getting translated to c int
what we want is unsigned long, so have to use 'long' which gets translated correctly to:
__Pyx_PyInt_From_unsigned_long(__pyx_v_xevent.xproperty.time);

+++ xpra/x11/bindings/window.pyx
@@ -1330,7 +1330,7 @@ cdef class X11WindowBindingsInstance(X11CoreBindingsInstance):
         return r==0
 
 
-    def get_server_time(self, Window xwindow) -> int:
+    def get_server_time(self, Window xwindow) -> long:
         cdef unsigned char c = b"a"
         cdef Atom timestamp_prop = self.xatom("XPRA_TIMESTAMP_PROP")
         XChangeProperty(self.display, xwindow, timestamp_prop,
